fix: move accordion components/table/
